fix nummericInput onChange bug
